refactor: extract Live flow into internal/llminternal/liveflow subpackage
Decompose base_flow_live.go (1,438 LOC, 54 functions, 6 concerns) into a
focused subpackage at internal/llminternal/liveflow/, matching the
existing convention for cohesive internal concerns (converters/,
googlellm/).

Files under internal/llminternal/liveflow/:
- flow.go               LiveFlow, RunLive, runSession, shared event channel
- retry.go              ConnectFn, reconnect knobs, GoAway handling
- sender.go             sender loop and queue draining
- receiver.go           receive worker, receiver loop, message processing
- turn_cycle_guard.go   turnCycleGuard + suppression matrix
- tools.go              tool coalescing, dispatch, execution, callbacks
- telemetry.go          liveTimingState, diagnostics builders

The three Live test files move with their production files.

Single caller-side change: agent/llmagent/llmagent.go swaps to
liveflow.LiveFlow / liveflow.ErrIter; the llminternal import remains
for non-Live model/tool callback types.

ErrIter (previously exported from llminternal) moves with the Live code
since it returns a Live-typed iterator. Two unexported symbols are
duplicated across the package boundary to avoid expanding API surface:
the pluginManager interface (tools.go) and the setupTestTracer /
testExporter test helpers (telemetry_test.go) — canonical Go pattern.

Pure structural refactor: no behavior change, no public-API rename.
Test count unchanged at 118; runner_live_test.go green.

Prepares for eventual upstream extraction to google/adk-go.
