+++ internal/llminternal/liveflow/turn_cycle_guard_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package llminternal
+package liveflow
 
 import (
 	"testing"
